uitests/requirements.txt
docopt~=0.6.2
behave~=1.2.6
requests~=2.21.0
selenium~=3.141.0
progress~=1.5
junitparser~=1.3.2
azure-cognitiveservices-vision-computervision
